requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-community
langchain-openai
chromadb
sentence-transformers
requests
pydantic
pytest
PyMuPDF
streamlit
tiktoken
tqdm
openai
